refactor: 모달 관련 로직을 modalView로 분리
